+++ linphone-app/ui/modules/Common/Picker/DatePicker.qml
@@ -5,6 +5,7 @@ import Common 1.0
 import Common.Styles 1.0
 import Linphone 1.0
 import Units 1.0
+import UtilsCpp 1.0
 
 import 'qrc:/ui/scripts/Utils/utils.js' as Utils
 
@@ -13,8 +14,9 @@ Item{
 	property bool allYears : false	// if false : years from today
 	property alias selectedDate: monthList.selectedDate
 	property bool hideOldDates : false
-	
-	signal clicked(date date);
+	property DateModel currentDate: UtilsCpp.getCurrentDateModel()
+	Component.onDestruction: {gc()}	// Call gc to free all DateModels
+	signal clicked(var date);
 	
 	RowLayout {
 		id: headerRow
@@ -36,7 +38,7 @@ Item{
 			Layout.fillWidth: true
 			Layout.alignment: Qt.AlignCenter
 			horizontalAlignment: Qt.AlignCenter
-			text: new Date(monthList.currentYear, monthList.currentMonth, 15).toLocaleString(App.locale, 'MMMM yyyy')// 15 because of timezones that can change the date for localeString
+			text: new Date(monthList.currentYear, monthList.currentMonth-1, 15).toLocaleString(App.locale, 'MMMM yyyy')// 15 because of timezones that can change the date for localeString
 			color: DatePickerStyle.title.colorModel.color
 			font.pointSize: DatePickerStyle.title.pointSize
 			font.capitalization: Font.Capitalize
@@ -60,23 +62,25 @@ Item{
 		property int maxYears: 5	// Max years to be requested.
 		
 		function set(date) {
-			selectedDate = new Date(date)
-			var moveTo = (selectedDate.getFullYear()-minYear) * 12 + selectedDate.getMonth()
-			currentIndex = moveTo
+			if(date) {
+				selectedDate = date
+				var moveTo = (selectedDate.year-minYear) * 12 + selectedDate.month - 1
+				currentIndex = moveTo
+			}
 		}
 		
-		property date selectedDate: new Date()
-		property int minYear: mainItem.allYears ? new Date(0,0,1).getFullYear() : new Date().getFullYear()
+		property DateModel selectedDate: mainItem.currentDate
+		property int minYear: (mainItem.allYears ? UtilsCpp.getMinDateModel() : mainItem.currentDate).year
 		
 		snapMode:    ListView.SnapOneItem
 		orientation: Qt.Horizontal
 		clip:        true
 		
 		// One model per month
-		model: (new Date().getFullYear()- minYear + maxYears) * 12
+		model: (mainItem.currentDate.year - minYear + maxYears) * 12
 		currentIndex: 0
 		property int currentYear:      Math.floor(currentIndex / 12) + minYear
-		property int currentMonth:     currentIndex % 12
+		property int currentMonth:     currentIndex % 12 + 1
 		
 		highlightFollowsCurrentItem: true
 		highlightRangeMode: ListView.StrictlyEnforceRange
@@ -89,8 +93,8 @@ Item{
 			height: monthList.height == 0 ? 100 : monthList.height
 			
 			property int year:      Math.floor(index / 12) + monthList.minYear 
-			property int month:     index % 12
-			property int firstDay:  new Date(year, month, 1).getDay()
+			property int month:     index % 12 + 1
+			property int firstDay:  UtilsCpp.getDateModel(year, month, 1).dayOfWeek();
 			
 			GridLayout { // 1 month calender
 				id: grid
@@ -109,12 +113,12 @@ Item{
 					
 					delegate: Item{
 						id: cellItem
-						property int day: index - 7 // 0 = top left below Sunday (-7 to 41)
-						property int date: day - firstDay + 1 // 1-31
-						property date cellDate: new Date(year, month, date)
+						property int dayIndex: index - 7 // 0 = top left below Sunday (-7 to 41)
+						property int date: dayIndex - firstDay + 1// 1-31
+						property DateModel cellDate: UtilsCpp.getDateModel(year, month, date)
 						property bool selected : text.text != '-' 
-											&& Utils.equalDate(cellDate, monthList.selectedDate)
-											&& text.text && day >= 0
+											&& text.text && dayIndex >= 0
+											&& cellDate.isEqual(monthList.selectedDate)	// Do not use == as it will compare JS Date.
 						width: grid.cellMinSize
 						height: width
 						
@@ -136,15 +140,15 @@ Item{
 								color: DatePickerStyle.cell.colorModel.color
 								font.pixelSize: cellItem.selected 
 													? DatePickerStyle.cell.selectedPointSize
-													: cellItem.day < 0 
+													: cellItem.dayIndex < 0
 														? DatePickerStyle.cell.dayHeaderPointSize
 														: DatePickerStyle.cell.dayPointSize
-								font.bold: cellItem.day < 0 || cellItem.selected || Utils.equalDate(cellItem.cellDate, new Date()) // today
+								font.bold: cellItem.dayIndex < 0 || cellItem.selected || (cellItem.cellDate && cellItem.cellDate.isEqual(mainItem.currentDate)) // today
 								text: {
-									if(cellItem.day < 0){
+									if(cellItem.dayIndex < 0){
 										// Magic date to set day names in this order : 'S', 'M', 'T', 'W', 'T', 'F', 'S' in Locale
 										return App.locale.dayName(index)[0].toUpperCase()
-									}else if(cellItem.cellDate.getMonth() == month && (!hideOldDates || new Date(year, month, cellItem.date+1) >= new Date()))	// new Date use time too
+									}else if(cellItem.cellDate && cellItem.cellDate.month == month && (!hideOldDates || cellItem.cellDate.isGreatherThan(currentDate)))
 										return cellItem.date
 									else
 										return '-'
@@ -156,7 +160,7 @@ Item{
 							id: mouseArea
 							
 							anchors.fill: parent
-							enabled:    text.text && text.text != '-' &&  cellItem.day >= 0
+							enabled:    text.text && text.text != '-' &&  cellItem.dayIndex >= 0
 							
 							onClicked: {
 								monthList.selectedDate = cellItem.cellDate