+++ CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## 5.1.3 - Undefined
 
 ### Fixed
+- Wrong dates from DatePicker.
 - Update SDK to 5.2.98
 
 ## 5.1.2 - 2023-08-25